Prevent panics when processing re-imports (i.e. `use` items).
Some callers of `FullyQualifiedName` expect to only work with named
items (e.g. `format_fn` expects that all functions it works with are
named; and `format_ty_for_cc` expects that all ADTs it works with are
named).  Before this CL `FullyQualifiedName::new` would panic when
called for a `def_id` without an associated name.  This panic would be a
problem for callers of `FullyQualifiedName` that only need the
`mod_path` (e.g. when `format_crate` emits unsupported-item comments and
uses `mod_path` to reorder them into the right location in relation to
other topo-sorted bindings).

This CL makes `FullyQualifiedName::name` an `Option` and tweaks
`FullyQualifiedName::new` so that it no longer panics when called with
an unnamed item.  This CL also opportunistically replaces the custom
`get_symbol` function with more-or-less equivalent calls to
[`DefPathData::get_opt_name`](https://doc.rust-lang.org/beta/nightly-rustc/rustc_hir/definitions/enum.DefPathData.html#method.get_opt_name).

PiperOrigin-RevId: 512684397