Fix sketchy code: don't use dangling pointers in string_view.
After this CL, there should be no UB from Rust/C++ interop on string_view/slices.

The circumstances where you can use Rust's dangling slice pointers in C++ are very limited. It should be possible to do this using `string_view(dangling, dangling)` but not `string_view(dangling, 0)`. (Because `[dangling, dangling)` forms a valid span, but `[dangling, dangling + 0)` does not, because `dangling + 0` is undefined behavior.)

How this is implemented in the standard library is opaque to us, and we are not invoking the constructors. So to avoid risk of UB, the safest thing to do for now is to not pass in dangling pointers.

The fix for now is to just branch. We could, later, try to avoid the branch by directly invoking the (it, it) constructor.

PiperOrigin-RevId: 512862276