Stop supporting `enum`s.
The main motivation is to limit the impact of ABI incompatibility from
b/270454629 (after this CL enums are unsupported and therefore are
not impacted by this bug).

This CL starts to differentiate between `enum`, `struct`, and `union`
types and therefore (in addition to removing `enum` support) starts
to translate Rust `union` into C++ `union` (before this CL all 3 kinds
of ADTs would be translated into an opaque C++ `struct`).

PiperOrigin-RevId: 513037796