Add `crubit_feature` aspect hints, without actually using them anywhere (yet).
Crubit features are a per-target configuration option which aim to alter how entities from these targets have their code generation. In particular, this allows for breaking changes to be rolled out via a feature rollout, and it allows for different targets to have access to different parts of Crubit. This CL introduces three sets of feature flags:

1. `[]`: the empty set, which will (in a followup CL) result in no bindings being generated.
2. `["supported"]`: stable features, which will (in a followup CL) result in some subset of Crubit functionality being enabled.
3. `["supported", "experimental"]`: all features enabled, but with visibility restricted.

These can be passed in by users using the following aspect hints:

1. `aspect_hints=[]` (the default)
2. `aspect_hints=["//third_party/crubit:supported"]`
3. `aspect_hints=["//third_party/crubit:experimental"]`

For rolling out a new backwards-incompatible Crubit change without a new binary release, we _could_ do something like:

1. Introduce a new "foo" feature and "foo_optout" feature, and check inside `src_code_gen.rs` whether `foo` is on or off. (off = either `foo_optout` is specified, or no value is specified.)
2. Run a TGP with `foo` added to `_SUPPORTED_FEATURES`.
3. For every C++ library that causes breakages in Rust targets, add `foo_optout` to that library.
4. Once `foo_optout` has been added to all relevant libraries, flip the default by adding `foo` to `_SUPPORTED_FEATURES` for real.
5. Fix each instance of `foo_optout`.
6. Once it's no longer used anywhere, delete the features entirely and perform the new behavior unconditionally.

At no point in this process do we break things with the binary release -- instead, it is only individual bazel changes (decoupled from the release) which can cause a large breakage and be quickly rolled back.

This is not the _only_ way to do it -- in particular, we could also use bazel flags to change defaults -- but the idea is, once we have this per-target configuration knob, we have the ability to perform these breaking changes in a manageable way, decoupled from the crubit binary release.

### targets_and_headers

Since this is essentially identical to how `targets_and_headers` works, I've merged it into the `targets_and_headers`, creating a new `target_args` flag. This includes all per-target arguments to Crubit, like what headers it has, and what features it has.

### Alternative Designs

Instead of using aggregated aspect hints which are exported in the provider via a depset, we _could_ have exported them in a file, as with namespaces. However, the depset approach is substantially simpler as it turns out. The only reason we need the namespace file is so that it can be read by the `cc_import` macro -- all other arguments can be passed in as a command line flag, as is done here.

Instead of using a function to return the features to the aspect implementation, we could have defined a _new_ aspect which returned it via a provider. This is easier to test, but now exposes a new provider, and idk probably performance or something. I wasn't sure if it was worth doing, so I didn't.

### Unresolved questions

... how do I actually permanently rename the old flag? I think I need to submit this CL, then in a separate CL rename the flag, and then once that's deployed, in a third CL, remove the alias, but I'm not totally sure. I am assuming it will work out.

PiperOrigin-RevId: 513039766