Remove non-Cpp20-aggregate trivial type from goldens.
The difference between aggregate and non-aggregate trivial types is not super important. If we _do_ want to test it, we should test it using a type which is a non-aggregate in both C++17 and C++20 and trivial for calls -- and we already have that, in crubit/rs_bindings_from_cc/test/golden/nontrivial_type.h (NontrivialUnpin).

As a final note, integration tests for structs with defaulted constructors already exist in e.g. crubit/rs_bindings_from_cc/test/struct/constructors/test.rs.

PiperOrigin-RevId: 513978101