Cleanup: Return `IR`, not `Rc<IR>`, from testing functions.
This makes it more obvious that you can mutate the IR for testing purposes. You could before, as well, using e.g. `Rc::make_mut(&mut ir)`, but it was clunky. And because the IR was behind an `Rc`, it wasn't obvious that it wasn't also stored somewhere else -- it could have been that you could mutate your copy, but there was some copy somewhere else that couldn't be mutated. That's what the API made it look like. By using `IR` by value, it's obvious that there's no shared state you'd be decoupling yourself from, and it's always OK to mutate the IR for testing purposes.

I think it was an `Rc<IR>` originally so that it could be passed in where an `Rc<IR>` is expected, but as this CL shows, it isn't hard to uncouple that.

PiperOrigin-RevId: 515266544