Use `supported | experimental` features everywhere in tests.
This actually was a bit tricky to implement, because it broke the formatting -- I discovered we use rustfmt for formatting the `Debug` output, which, while it mkaes sense, seems probably unnecessary if we do some plumbing. But I couldn't figure out the plumbing right now, so I filed b/272530008 instead.

After reflecting for this bug for a bit, I decided the sensible thing is to try to share the logic for both in one helper function, even if it seems a little awkward to do so -- having _no_ shared functionality is essentially what caused it.

PiperOrigin-RevId: 516681109