Handle colon-less and relative bazel labels in `ir.rs`.
This has been bugging me for a while now -- not because I'd predict it to be a real issue, but just because it feels weird to rely on a specific formatting of bazel labels. This has bitten me in the past with e.g. unknown commit.

I restrained myself until now, but it's still kind of bothering me that we crash on `//foo` syntax. I took the most expedient option of fixing it the simplest way I know how.

The alternative would have been to make the field private, and validate the input at parsing time (e.g. defining `TryFrom` instead of `From`, and validating that the label meets basic validity checks. I chose not to because it's a lot more work than to accept invalid labels without crashing.

This probably won't actually affect anything, but if we ever find ourselves forwarding bazel labels along using non-canonicalized form of `//foo:bar`, we will DTRT. Hopefully that doesn't hide any problems.

PiperOrigin-RevId: 516897644