Make a common `Item` -> `UnsupportedItem` path.
This removes the myriad ways to create an `UnsupportedItem`, and will make it easier to convert an item to "unsupported" for new reasons, as well. (In particular, I want to convert arbitrary items to unsupported in a followup CL if they are filtered out by the features flags checks.)

I was tempted to use a custom type implementing `Display`, rather than just `Rc<str>`, but it's actually an annoying amount of work and doesn't hide the type completely. (You end up with `type DebugName<'a> : Display; fn debug_name<'a>(&'a self, ir: &'a IR) -> Self::DebugName<'a>;` so that the Display impl can hold onto the two references.)

PiperOrigin-RevId: 518732262