rs_bindings_from_cc: Emit `arbitrary_self_types` feature flag only when it's needed
Previously, I added the feature flag unconditionally, unaware that there had been infrastructure for conditionally emitting feature flags already.

The `arbitrary_self_types` feature flag is only needed for bindings of ref-qualified methods: for an rvalue-ref-qualified method, the generated binding takes `RvalueReference` as `self`, and the flag is needed to allow it.

PiperOrigin-RevId: 519993346