Forbid putting lifetime annotations on pointers or references to functions.
This also adds an end-to-end test in rs_bindings_from_cc.

Alternative considered:

I explored the alternative approach of eliminating the function pointer entirely
from the tree of `ValueLifetimes`. This would be done by doing the following
in `ValueLifetimes::Create()`:

```
if (type->isFunctionPointerType() || type->isFunctionReferenceType()) {
  return value_lifetimes;
}
```

This initially seemed to have the advantage of more directly expressing the
logical intent that "function pointers aren't pointers". However, it had the
undesirable consequence of breaking the invariant
`ValueLifetimes::Create(type).Type() == type`. This in turn broke the invariant
that an `Object` has the same type as the `Expr` it is associated with in the
`PointsToMap`. This necessitated a lot of special-casing in
`TransferStmtVisitor` and seemed like a much worse approach overall.

PiperOrigin-RevId: 520621392