Remove remaining TODOs related to `'static` lifetime of `fn` pointers.
After the recent work on b/275628345 the `lifetime_annotations` layer
guarantees that function pointers and references either 1) have no
associated lifetime (without lifetime elision) or 2) have an inferred
`'static` lifetime.  This CL tweaks the associated TODOs and assertions
in `importer.cc` and `src_code_gen.rs` to reconcile that with that work.

(Note that the text of the TODOs has incorrectly assumed that there
will be always no lifetime associated with function pointers/references.
The CL changes the assertions and comments to agree with the current
behavior in this area.)

The CL also removes the `test_func_ptr_with_non_static_lifetime` test.
The test is no longer applicable (since we reject function pointers
with non-`'static` lifetime in the `lifetime_annotations` layer (*).
Also note that the test used a no-op assertion: an `assert_rs_matches`
with an empty `pattern` (removal of the test is necessary to move
forward with unknown commit).

(*) I don't understand why the unit test in `src_code_gen.rs` doesn't
report any compile errors when `$a` lifetime is used in the unit test.
This might be interesting to investigate, but doesn't change the
need to remove the obsolete test.

PiperOrigin-RevId: 521468722