Don't ignore all errors from `GetLifetimeAnnotations()`.
So far, if `GetLifetimeAnnotations()` returned an error, we simply ignored it
and proceeded without lifetimes (i.e. importing C++ pointers and references as
Rust raw pointers). Implicitly, this assumed that the error was of the type
"lifetime elision not enabled" or "lifetime elision not applicable".

However, `GetLifetimeAnnotations()` can also return other types of error, and
these should be treated as hard errors; the API should not be imported in this
case. The test added here demonstrates on such case.

`GetLifetimeAnnotations()` has been changed to return structured error
information, so we can now distinguish between the different types of error and
treat them as applicable.

PiperOrigin-RevId: 522259997