Support ZST fields in bindings generation.
To pick a canonical order when multiple ZSTs are in the same memory location, we pick the order they were listed in source code, which makes sense enough.

To actually represent them, the code as-is would've produced an `unsigned char field[0]`, which is ill-formed but often implementation-defined. I'm not sure if it's that much better, but for now I replaced them with `[[no_unique_address]] struct{} field;`.

This does have the clear advantage that, unlike `0`-sized arrays, it is guaranteed to work on all conforming C++20 implementations, though it does require C++20. It's also probably similar to how we intend to implement ZSTs anyway -- it seems close to certain that ZSTs will "just" be empty structs which have `[[no_unique_address]]` whenever they're used as a field. In that case, this is a stepping stone towards that future: now "all" we need to do is to replace the anonymous struct with the real ZST struct.

PiperOrigin-RevId: 522418138