Convert `&mut dyn ErrorReporting` into (interior mutable) `Rc<dyn ErrorReporting>`, available globally in `db`.
The motivating use case here is to allow it to be a parameter inside a `db` method: to be a parameter, it needs to be `Debug` and cheaply clonable. However, parameters also require equality-comparison. Since this parameter is fixed for the whole run, it makes more sense (and is less code) to make it a whole salsa input.

The reason to allow it as a parameter / used inside a `db` method is to allow these to call `generate_item` as a reverse edge. I need this in a successor CL to assert bindings exist. For example, an alias declaration should not be generated unless the type it aliases is successfully generated.

PiperOrigin-RevId: 523773547