Migrate `cc_bindings_from_rs` from a rule, to an aspect with a wrapper rule.
This wrapper rule takes the place of `cc_deps` in rules rust -- it grants access to the `CcInfo` provider that was generated by Crubit, which would otherwise be inaccessible. I'm not sure of any alternative way of exposing this, short of actually _modifying_ `cc_binary` to, firstly, accept our `CcBindingsFromRustInfo` in addition to `CcInfo`, and, secondly, follow our aspect. I don't think that making it work without a wrapper rule is necessary to mark b/270754283 as fixed -- so far as I know, this CL moves the logic into an aspect in a satisfactory way / which allows us to start doing automatic depset stuff, so I've marked the CL fixed in this CL.

This change necessary in order to be able to automatically compute the location of generated header files: because the generated header files are located in one canonical place, from the aspect, we can now have things like cross-crate dependencies that work automatically.

The major consequence visible in this CL is that generated bindings occur next to the _wrapped library_, and not next to the _wrapper rule_. The best place to see this is in the crates_io test. This means that visibility of utility libraries must be expanded to _all_ packages, and so on. This is unavoidable, of course.

I have also helpfully provided a `OutputGroupInfo`, just like `rs_bindings_from_cc`. This means that we can e.g. do a similar trick as b/273803618 in a followup CL, to review diffs of generated files. It also means we can directly read the files on the command line by specifically requesting them. For example:

```
bazel build --aspects //third_party/crubit/cc_bindings_from_rs/bazel_support:cc_bindings_from_rust_rule.bzl%cc_bindings_from_rust_aspect --output_groups=out //third_party/rust/fastrand/v1:fastrand
```

---

P.S. this completely shattered the functionality of the bazel unit tests. I could not figure out how to fix them: moving the logic into an aspect seems to remove it from `analysistest.target_actions()`. Since the test seems to be excessively tightly coupled to the implementation, and didn't catch any errors not caught by the other tests, and is, moreover, really difficult to fix from my POV -- I felt like maybe I could sensibly propose deleting the tests.

PiperOrigin-RevId: 523811708