Implement "soft" failures for crubit features, but don't use them yet.
Before this CL, failures to generate bindings for an item are almost always treated as a hard error,
because bindings generated in other translation units which (transitively) depend on this one can't
know that it failed. Function items were hardcoded as an exception to this.

After this CL, in addition to function items, any disabled item also is allowed to fail. A function
exists, `has_bindings`, which returns whether or not an item is guaranteed to be present.
(It returns `Maybe` for functions and for disabled items.)

Followup CL(s?) will check this flag when creating things which depend on these soft-disabled items. For example, when creating a type alias, we should check if `has_bindings(type)`, and only generate bindings for the _alias_ if the aliased type itself is guaranteed to receive bindings.

---

My intent originally had been to actually use `generate_item` for this, but as you'd probably have guessed faster than me, we can't really do this after all, because this forms a cyclic dependency. Suppose we want to write bindings for `struct S {S* x;};` -- we need to know, while generating bindings for the `x` field, if `S` will get bindings, which we are doing in the very process of creating bindings for `S`.

The "good" solution might be to have cycle handling, so that if we detect a cycle, we allege that bindings generation will succeed, since the answer is either `true` or irrelevant. But we're on a 2y old version of salsa (it hasn't received any updates), and my understanding was that the next version of salsa will revamp cycle handling. Also, it's pretty gnarly and I don't actually know how to do it in Salsa.

Easier is to move the logic out. I admit this is not as pretty in some ways, but it certainly works and doesn't require fancy cycle detection.

(It's also faster, since we won't need to run bindings generation for all our transitive dependencies -- we can just call `has_bindings`. And really, speed is nice.)

PiperOrigin-RevId: 524104811