Handle aliases (and all sugar) in type => nullability vector computation
When walking a type looking for nullability slots, we assume an unhandled type
node has no slots. We weren't handling type aliases, so we'd compute an empty
nullability vector for `T` in `using T = int*;`, oops.

Instead of just handling aliases specifically, add fallback handling for all
sugar types to walk the underlying type.
We can remove the uninteresting specific handling for certain sugar types.
The special behavior for AttributedType still works, of course!

Also added unittest covering getNullabilityAnnotationsFromType, which is the
simplest way to test this change.
This doesn't *nearly* cover all interesting cases, but will at least make it
easier to add more later/migrate some away from __assert_nullability/test
subsequent changes to nullability vector computation.

This bug was the root cause of a problem identified with
https://github.com/google/crubit/commit/7a808ea25d183f5f788c0fb3342ffec99d16c829
When handling a cast from nullopt to a pointer type, that change assumes the
nullability vector of the result type is nonempty (it's a pointer type!) and
assigns to the front of it. But when aliases were involved we got an incorrect
empty vector.

PiperOrigin-RevId: 524813522