Disable tests that can't run on ARM itself.
These tests, by and large, fail because they depend on `clang-format`. Note that `clang-format` is a build-time dependency, and thus we only ever use it from `x86`. So, strictly speaking, these tests _do not need to run on ARM_. Instead, they can run on x86, and where they intersect with ARM, it's only via cross-compilation, and we should forward the appropriate flags to clang/etc. to make it target an arm triple.

Maybe this is optional and we can just never run these tests, instead relying exclusively on integration tests to assure us that Crubit works on ARM. I'm not really sure that's a good idea, though. :)

It is probably especially important to do this kind of parameterization when we add support for the Windows ABI, which is much trickier than just Itanium-ARM.

PiperOrigin-RevId: 525843912