Improve interpretation of nullability in type aliases, and uncover a new bug.
Aliases are not opaque, and the nullability written on the alias is not the
right nullability for the underlying canonical type.
Walking the underlying type is an improvement...

...but not quite sufficient: where template params are referenced in the
alias template defn, they are replaced by SubstTemplateTypeParmType which
desugars to the *canonicalized* arg.
In other words, we lose sugar when type alias args are expanded.
It's recoverable (similar to how template substitutions work in e.g.
MemberExpr) but it's fiddly so just add a failing test for now.

PiperOrigin-RevId: 526952363