Don't discard sugar between the _Nullable attribute and the PointerType.
It's valid to place this attribute on any type that is *canonically* a pointer.
Before this CL, we eagerly unwrap to find the PointerType to associate with.
The sugar we unwrap and discard may be significant, e.g.:
  template <class T> struct Nullable { using type = T _Nullable; };
  using Target = Nullable<int * _Nonnull *>::type;
Here _Nullable applies to a TemplateSubstTypeParmType. getAs<PointerType>()
returns its underlying type int**, and we have no hope of recovering the inner
_Nonnull after that.
However walking the sugar normally, we can find the TemplateSubstTypeParmType,
identify it with the template arg found in the ElaboratedType, and walk
`int *_Nonnull *` instead, attaching _Nullable to the outer pointer type and
still finding the inner nullability.

Unfortunately no new test for this: we don't yet handle such template cases.
I can't find a case where the sugar we discard is significant.
Existing sugar tests pass though.
(I'm working on a patch to handle template aliases, this is a prerequisite)

PiperOrigin-RevId: 527270901