Test host-only code with cross-compilation.
Since Crubit itself only is expected to run on x86 (right now?), cross-compiling to arm, it makes sense to do the same in the unit tests. This makes the unit tests more like production, and avoids needing to actively work to get the rest of the internals working on arm. It's possible, but nah.

This was constructed by parameterizing the tests using an environment variable. This was easier than parameterizing the tests using loops, rstest, etc., since those would require adding each parameter to each function, when really we want to do ~all parameters for these (presumably).

This testing strategy only really works because both arm-linux and x86-linux use the Itanium ABI, which is effectively identical. If/when we add non-Itanium support, we'd need to add separate files, and restrict the tests to e.g. only run on Itanium platforms.

There is one major integration-y test that doesn't cross-compile for Arm, generate_bindings_and_metadata_test, but I think it's of relatively low value given the remaining "unit" tests and the integration tests. I'm fairly comfortable saying we test Arm -- we may even test it _too_ much, given how similar the two platforms are. We can mark the bug as fixed: Arm is tested and works.

PiperOrigin-RevId: 527324587