Resugar substituted template type params in type alias templates
a.k.a. make `Nullable<Nullable<int*>*>` work.

When we enter a type alias specialization, record the sugared arg list.
When we see a substituted alias param within it look up the corresponding arg,
and walk that instead of the underlying type (which is canonicalized).
The substituted param always corresponds to the innermost alias, as aliases
can't be nested so it's the only set of params available.

While here, add a test for a common (non-alias) template case we don't handle...

PiperOrigin-RevId: 527510439