Function to print a type and its nullability vector together in standard syntax
This will be used in printing test failures, in the analysis logs.
(Potentially in diagnosis one day?)

It further decouples us from the nullability vector data structure.

The main tradeoff is complexity: this is a third algorithm (after transfer &
NullabilityWalker) that matches the data structure to AST nodes, and the three
must be kept in sync.

PiperOrigin-RevId: 527960895