Split functions relating to type-nullability out of pointer_nullability.h
The former has been kind of an "everything that needs to be shared" grab-bag,
and it's growing...
The functionality here today splits neatly into two groups, one that interacts
with the Value world and one that interacts with the Type world, with analysis
depending on both.

The exception is getNullabilityForChild, which fits naturally into analysis.
It was moved here with the plan of being shared with diagnosis. but hasn't been.
It doesn't seem suitable to sharein its current form - it causes other layering
problems too, and its preconditions/API are defined in terms of analysis.
So I've moved this back into analysis for now.

Wrote a few words about the type-nullability model.
I'd like to follow up by renaming a few functions, adding a TypeNullability
alias, adding more tests, but trying to keep the scope for discussion smaller.

PiperOrigin-RevId: 528755134