[LifetimeAnalysis] Fix potential breakages for PruneTriviallyFalseEdges
Change some tests not to use the constant boolean value for conditionals
so that the analysis (correctly) takes care of the both branches even if
unreachable nodes are pruned.

Also make a few fixes not to crash when PruneTriviallyFalseEdges is set
to true.

This is a preparation for landing https://reviews.llvm.org/D149640

PiperOrigin-RevId: 529029325