Allow for `union`s as record members
When `ObjectLifetimes::GetObjectLifetimesForTypeInContext` gets called
for record members, it was skipping `union`s because the condition for
analyzing record-like types was `type->isStructureOrClassType()` instead
of `type->isRecordType()`. This caused the default case to create a
"liftime-less" type with `ForLifetimeLessType`. However, the passed type
is a `union` which causes the following assertion to fire:

`PointeeType(type).isNull() && !type->isRecordType())`

PiperOrigin-RevId: 529072296