[nullability] Avoid crashing on alias templates with default arguments.
If an alias template has default arguments, the `SubstTemplateTypeParmType` may
not have an associated argument in
`TemplateSpecializationType::template_arguments()`. We used to crash on this
with an out-of-bounds access.

This patch adds a test that triggers this condition. For the time being, we
merely avoid crashing, but longer-term we want to actually access the default
argument and resugar it; see comments in the code for details.

PiperOrigin-RevId: 530242561