Implement `Display` on RsTypeKind.
This requires exporting a non-rustfmt-invoking serialization for TokenStreams.

There's a bit of an open question on what to do with broken `__COMMENT__` tokens. I can't help but wonder: maybe we shou;d just, uh, not fail on them? For the moment, though, it's easy enough to just abort partway through serialization and mark it as broken.

PiperOrigin-RevId: 532187123