Emit debug size/align checks for type map overrides.
This is not an in-depth guarantee of correctness, but it does catch obvious mistakes.

Mistakes it does not catch: mismatch in C ABI when calling by value; mismatch in field offsets; breakage of internal invariants due to incorrect actual values (e.g. std::vector with size > capacity, that sort of thing).

PiperOrigin-RevId: 532188571