Fix alignment of the generated structs that contain bitfields.
The alignment may be wrong when replacing a bitfield aligned at N bytes
with an opaque padding (blob of bytes has no alignment).  In some cases
presence of other fields may still result in the right alignment, but
this requires that the other fields are 1) present and 2) have the same
or higher alignment as the replaced bitfield.

PiperOrigin-RevId: 535673317