Generate bindings for `using foo::SomeType` as if it were `using SomeType = foo::SomeType`.
This is not as obvious as it seems. For example, Kythe doesn't do this, it skips right over using-aliases, producing somewhat confusing mouseovers. This is because, well, a using declaration **is not a type** in the clang AST, it's just a name declaration.

The consequence of that is that a bunch of things that used to strictly require types now accept `NamedDecl` objects, so that we can pass in a `using` decl.

I don't think this is totally outrageous, but it is unfortunate.

Anyway, the end result is that `using` does pretty much what you'd guess it does, when you use it in a header file.

It wasn't super important that I fix this, but existing tests for `rs_char` used `using rs_char::rs_char` in a header, and I wanted to keep that working when I changed `rs_char` to be a little less "hardcoded". Thus, this CL.

PiperOrigin-RevId: 535735712