One half of bidirectional interop: `RustBindings(CcBindings(X)) == X`.
This replaces both dependencies and types, so that all uses of the bindings use the original Rust library, and all uses of a Rust type use the original Rust type.

This is not entirely complete, in that if we specified CRUBIT_INTERNAL_FORCE_C_ABI_COMPATIBLE, it would break, because bindings generation currently assumes that C-ABI-compatible types are copyable, etc. But I think that's better left to the more general bug about C-ABI-compatible types.

---

The bazel stuff is ever so slightly messy -- I think perhaps CcBindingsFromRustInfo should be split into two providers, one which gives the Rust dep info, and one which gives bindings/C++ info. But I wasn't sure the best way to do this since it's also used from third_party.

Note: an alternate approach would've been to _add_ a dependency on the original Rust crate from its bindings. This works (thanks to `force_all_deps_direct=True`), but still has an unnecessary C++ library involved. The current version of the CL short-circuits bindings generation entirely.

---

Followup work: do exactly all this but in reverse, for cc_bindings_from_rs.

PiperOrigin-RevId: 536552702